Share the remembered-session key and parse shape across bundle trees
